core: move read transitions into their own module
