+++ core/src/world_ops.rs
@@ -1,6 +1,6 @@
 //! Protocol-neutral world transitions.
 //!
-//! This module owns disk-side read / replace / append semantics:
+//! This module owns disk-side replace / append semantics:
 //! auth permits, body caps, per-world locks, tombstone clearing,
 //! preconditions, quota reservation, durable+memory writes, audit,
 //! and notify. It deliberately returns semantic outcomes instead of wire
@@ -14,53 +14,17 @@ use std::sync::atomic::Ordering;
 use bytes::Bytes;
 
 use crate::{
-    audit, auth, can_read, can_write,
+    audit, auth, can_write,
     engine_types::{ChangeVerb, ValidatedWorldPath},
     etag, needs_write_approve, store,
-    timeline::{BodySha256, TimelineAddress, TimelineRead},
+    timeline::BodySha256,
     world, AuthGate, Core, StorageFailureClass,
 };
-
-#[derive(Debug)]
-pub(crate) struct ReadPermit {
-    world: ValidatedWorldPath,
-}
 
 #[derive(Debug)]
 pub(crate) struct WritePermit {
     world: ValidatedWorldPath,
     gate: AuthGate,
-}
-
-pub(crate) enum ReadOutcome {
-    Found { stage: world::Stage, etag: String },
-    Missing,
-}
-
-#[derive(Debug)]
-pub(crate) enum ReadError {
-    Auth(AuthGate),
-    TransientStorage {
-        #[allow(dead_code)]
-        scope: &'static str,
-        err: rusqlite::Error,
-    },
-    InsufficientStorage {
-        #[allow(dead_code)]
-        scope: &'static str,
-        err: rusqlite::Error,
-    },
-    StorageRead {
-        #[allow(dead_code)]
-        scope: &'static str,
-        err: rusqlite::Error,
-    },
-    AuditChainBroken {
-        #[allow(dead_code)]
-        scope: &'static str,
-        break_report: audit::VerifyBreak,
-    },
-    PermitWorldMismatch,
 }
 
 #[derive(Debug)]
@@ -157,20 +121,6 @@ pub(crate) trait WriteTraceHooks {
     fn notify_sent(&self) {}
 }
 
-pub(crate) fn authorize_read(
-    core: &Core,
-    world: &ValidatedWorldPath,
-    tier: auth::Tier,
-) -> Result<ReadPermit, ReadError> {
-    if can_read(core, tier) {
-        Ok(ReadPermit {
-            world: world.clone(),
-        })
-    } else {
-        Err(ReadError::Auth(AuthGate::Read))
-    }
-}
-
 pub(crate) fn authorize_write(
     world: &ValidatedWorldPath,
     tier: auth::Tier,
@@ -187,42 +137,6 @@ pub(crate) fn authorize_write(
         })
     } else {
         Err(WriteError::Auth(gate))
-    }
-}
-
-pub(crate) fn read_world(core: &Core, permit: &ReadPermit) -> Result<ReadOutcome, ReadError> {
-    read_world_for(core, permit, &permit.world)
-}
-
-pub(crate) fn read_timeline_body(
-    core: &Core,
-    permit: &ReadPermit,
-    address: &TimelineAddress,
-) -> Result<TimelineRead, ReadError> {
-    if permit.world != *address.world() {
-        return Err(ReadError::PermitWorldMismatch);
-    }
-    match core.read_timeline_body(address) {
-        Ok(Some(read)) => Ok(read),
-        Ok(None) => Ok(TimelineRead::Gone {
-            address: address.clone(),
-        }),
-        Err(err) => Err(classify_read_audit_error("timeline read", err)),
-    }
-}
-
-pub(crate) fn read_world_for(
-    core: &Core,
-    permit: &ReadPermit,
-    world: &ValidatedWorldPath,
-) -> Result<ReadOutcome, ReadError> {
-    if &permit.world != world {
-        return Err(ReadError::PermitWorldMismatch);
-    }
-    match core.read_world_with_etag(world.as_str()) {
-        Ok(Some((stage, etag))) => Ok(ReadOutcome::Found { stage, etag }),
-        Ok(None) => Ok(ReadOutcome::Missing),
-        Err(err) => Err(classify_read_error("storage read", err)),
     }
 }
 
@@ -474,24 +388,6 @@ fn check_write_preconditions(
         .map_err(|message| WriteError::PreconditionFailed { message })
 }
 
-fn classify_read_error(scope: &'static str, err: rusqlite::Error) -> ReadError {
-    match crate::classify_storage_failure(&err) {
-        StorageFailureClass::InsufficientStorage => ReadError::InsufficientStorage { scope, err },
-        StorageFailureClass::Transient => ReadError::TransientStorage { scope, err },
-        StorageFailureClass::Other => ReadError::StorageRead { scope, err },
-    }
-}
-
-fn classify_read_audit_error(scope: &'static str, err: audit::AuditError) -> ReadError {
-    match err {
-        audit::AuditError::ChainBroken(break_report) => ReadError::AuditChainBroken {
-            scope,
-            break_report,
-        },
-        audit::AuditError::Storage(err) => classify_read_error(scope, err),
-    }
-}
-
 fn classify_write_audit_error(scope: &'static str, err: audit::AuditError) -> WriteError {
     match err {
         audit::AuditError::ChainBroken(break_report) => WriteError::AuditChainBroken {
@@ -525,22 +421,9 @@ fn classify_write_storage_error(
 mod tests {
     use super::*;
     use crate::test_support::test_core;
-    use crate::{
-        timeline::{TimelineAddress, TimelineSeq},
-        world_generation::WorldGeneration,
-    };
 
     fn world_path(world: &str) -> ValidatedWorldPath {
         ValidatedWorldPath::new(world).unwrap()
-    }
-
-    fn timeline_address(world: ValidatedWorldPath) -> TimelineAddress {
-        TimelineAddress::test_only_new(
-            world,
-            WorldGeneration::new("0123456789abcdef0123456789abcdef").unwrap(),
-            TimelineSeq::new(1).unwrap(),
-            BodySha256::for_body(b"value"),
-        )
     }
 
     #[tokio::test]
@@ -569,64 +452,6 @@ mod tests {
         );
         assert!(core.read_world("home/permit-b").unwrap().is_none());
 
-        let _ = std::fs::remove_dir_all(dir);
-    }
-
-    #[test]
-    fn timeline_read_permit_is_bound_to_address_world() {
-        let (core, dir) = test_core("timeline-read-permit-bound");
-        let permit = authorize_read(&core, &world_path("home/right"), auth::Tier::Read).unwrap();
-        let address = timeline_address(world_path("home/wrong"));
-
-        assert!(matches!(
-            read_timeline_body(&core, &permit, &address),
-            Err(ReadError::PermitWorldMismatch)
-        ));
-
-        let _ = std::fs::remove_dir_all(dir);
-    }
-
-    #[tokio::test]
-    async fn real_sqlite_lock_maps_to_transient_read_error() {
-        struct NoopTrace;
-        impl WriteTraceHooks for NoopTrace {}
-
-        let (core, dir) = test_core("runtime-lock-classification");
-        let world = world_path("home/busy");
-        let write_permit = authorize_write(&world, auth::Tier::Write).unwrap();
-        replace_write(
-            &core,
-            &write_permit,
-            ReplaceRequest {
-                body: Bytes::from_static(b"ok"),
-                content_type: "text/plain".to_owned(),
-                headers: Vec::new(),
-                preconditions: etag::Preconditions::default(),
-            },
-            &NoopTrace,
-        )
-        .await
-        .unwrap();
-
-        let holder = rusqlite::Connection::open(crate::world::world_db(&dir, world.as_str()))
-            .expect("open lock holder");
-        holder
-            .pragma_update(None, "locking_mode", "EXCLUSIVE")
-            .expect("exclusive locking mode");
-        holder
-            .execute_batch("BEGIN EXCLUSIVE")
-            .expect("hold exclusive transaction");
-
-        let read_permit = authorize_read(&core, &world, auth::Tier::Read).unwrap();
-        assert!(
-            matches!(
-                read_world(&core, &read_permit),
-                Err(ReadError::TransientStorage { .. })
-            ),
-            "real SQLite busy/locked errors must stay classified as transient"
-        );
-
-        drop(holder);
         let _ = std::fs::remove_dir_all(dir);
     }
